requirements.txt
Flask==2.3.2
twilio==8.0.0
requests==2.31.0
python-dotenv==1.0.0
APScheduler==3.10.1
SQLAlchemy==2.0.22
gunicorn==20.1.0
openai==1.7.2

